requirements.txt
# Apache Airflow with PostgreSQL support
apache-airflow[postgres]==3.0.2

# Database Providers
apache-airflow-providers-postgres>=5.12.0

# Database Connectivity
pg8000>=1.30.0
psycopg2-binary>=2.9.0


